6/11

+++ src/app/student-delete/student-delete.component.ts
@@ -1,6 +1,9 @@
+// student-delete.component.ts
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { StudentService } from '../service/student.service';
+import { catchError } from 'rxjs/operators';
+import { of } from 'rxjs';
 
 @Component({
   selector: 'app-student-delete',
@@ -9,6 +12,8 @@ import { StudentService } from '../service/student.service';
 })
 export class StudentDeleteComponent implements OnInit {
   id: number = 0;
+  errorMessage: string | null = null;
+
   constructor(
     private route: ActivatedRoute,
     private router: Router,
@@ -17,13 +22,26 @@ export class StudentDeleteComponent implements OnInit {
 
   ngOnInit(): void {
     this.id = this.route.snapshot.params['id'];
-    this.deleteStudent();
-  }
-
-  deleteStudent():void{
-    this.studentService.deleteStudent(this.id).subscribe(() => {
+    if (confirm('Are you sure you want to delete this student?')) {
+      this.deleteStudent();
+    } else {
       this.router.navigate(['/students']);
-    });
+    }
   }
 
+  deleteStudent(): void {
+    this.studentService.deleteStudent(this.id)
+      .pipe(
+        catchError(error => {
+          this.errorMessage = 'There was an error deleting the student. Please try again.';
+          return of(null); // Return a null observable
+        })
+      )
+      .subscribe(response => {
+        if (response !== null) {
+          alert(response);
+          this.router.navigate(['/students']);
+        }
+      });
+  }
 }
